Organise types
