+++ apps/govern/types/Allocation.ts
@@ -1,0 +1,4 @@
+import { Address } from './Address';
+import { Metadata } from './Contract';
+
+export type Allocation = { address: Address; chainId: number; metadata: Metadata; weight: number };
