+++ apps/govern/types/Vote.ts
@@ -1,0 +1,7 @@
+export type UserVote = {
+  slope: number;
+  power: number;
+  end: number;
+};
+
+export type UserVotes = { current: UserVote; next: UserVote };
