+++ apps/govern/types/index.ts
@@ -1,1 +1,4 @@
 export type * from './Address';
+export type * from './Allocation';
+export type * from './Contract';
+export type * from './Vote';
